+++ common/src/main/java/com/github/teamfusion/spyglassplus/mixin/exclusive/client/ControlsListWidgetKeyBindingEntryMixin.java
@@ -1,7 +1,8 @@
-package com.github.teamfusion.spyglassplus.mixin.client;
+package com.github.teamfusion.spyglassplus.mixin.exclusive.client;
 
 import com.github.teamfusion.spyglassplus.client.keybinding.SpyglassPlusKeyBindings;
 import com.github.teamfusion.spyglassplus.mixin.client.access.KeyBindingAccessor;
+import com.github.teamfusion.spyglassplus.mixin.exclusive.client.access.ExclusiveKeyBindingAccessor;
 import net.fabricmc.api.EnvType;
 import net.fabricmc.api.Environment;
 import net.minecraft.client.gui.screen.option.ControlsListWidget;
@@ -39,7 +40,7 @@ public class ControlsListWidgetKeyBindingEntryMixin {
     private void onRender(MatrixStack matrices, int index, int y, int x, int entryWidth, int entryHeight, int mouseX, int mouseY, boolean hovered, float tickDelta, CallbackInfo ci, boolean selected, boolean incompatible) {
         if (incompatible) {
             if (this.binding == SpyglassPlusKeyBindings.TRIGGER_COMMAND_ENCHANTMENT
-                || this.binding == KeyBindingAccessor.getKEY_TO_BINDINGS().get(((KeyBindingAccessor) this.binding).getBoundKey())
+                || this.binding == ExclusiveKeyBindingAccessor.getKEY_TO_BINDINGS().get(((KeyBindingAccessor) this.binding).getBoundKey())
             ) this.editButton.setMessage(this.editButton.getMessage().copy().formatted(Formatting.WHITE));
         }
     }
