+++ common/src/main/java/com/github/teamfusion/spyglassplus/mixin/exclusive/client/access/ExclusiveKeyBindingAccessor.java
@@ -1,4 +1,4 @@
-package com.github.teamfusion.spyglassplus.mixin.client.access;
+package com.github.teamfusion.spyglassplus.mixin.exclusive.client.access;
 
 import net.fabricmc.api.EnvType;
 import net.fabricmc.api.Environment;
@@ -7,11 +7,10 @@ import net.minecraft.client.util.InputUtil;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.gen.Accessor;
 
+import java.util.Map;
+
 @Environment(EnvType.CLIENT)
 @Mixin(KeyBinding.class)
-public interface KeyBindingAccessor {
-    @Accessor InputUtil.Key getBoundKey();
-
-    @Accessor void setTimesPressed(int timesPressed);
-    @Accessor int getTimesPressed();
+public interface ExclusiveKeyBindingAccessor {
+    @Accessor static Map<InputUtil.Key, KeyBinding> getKEY_TO_BINDINGS() { throw new AssertionError(); }
 }
