+++ common/src/main/resources/spyglassplus-common.mixins.json
@@ -15,27 +15,12 @@
     "access.ThreadedAnvilChunkStorageAccessor"
   ],
   "client": [
-    "client.BipedEntityModelMixin",
-    "client.CameraMixin",
-    "client.ClientPlayerEntityMixin",
-    "client.ClientPlayerInteractionManagerMixin",
-    "client.ControlsListWidgetKeyBindingEntryMixin",
-    "client.EndCrystalEntityMixin",
-    "client.EntityMixin",
-    "client.EntityRendererMixin",
-    "client.GameRendererMixin",
-    "client.InGameHudMixin",
-    "client.ItemModelsMixin",
-    "client.KeyBindingMixin",
-    "client.LightmapTextureManagerMixin",
-    "client.LivingEntityMixin",
-    "client.LivingEntityRendererMixin",
-    "client.MinecraftClientMixin",
-    "client.ModelPredicateProviderRegistryMixin",
-    "client.MouseMixin",
-    "client.PlayerHeldItemFeatureRendererMixin",
-    "client.WorldRendererMixin",
-    "client.access.KeyBindingAccessor",
+    "client.BipedEntityModelMixin", "client.CameraMixin", "client.ClientPlayerEntityMixin",
+    "client.ClientPlayerInteractionManagerMixin", "client.EndCrystalEntityMixin", "client.EntityMixin",
+    "client.EntityRendererMixin", "client.GameRendererMixin", "client.InGameHudMixin", "client.ItemModelsMixin",
+    "client.LightmapTextureManagerMixin", "client.LivingEntityMixin", "client.LivingEntityRendererMixin",
+    "client.MinecraftClientMixin", "client.ModelPredicateProviderRegistryMixin", "client.MouseMixin",
+    "client.PlayerHeldItemFeatureRendererMixin", "client.WorldRendererMixin", "client.access.KeyBindingAccessor",
     "client.access.KeyBindingInvoker"
   ],
   "injectors": {
