+++ fabric/src/main/resources/spyglassplus-exclusive.mixins.json
@@ -3,9 +3,12 @@
   "package": "com.github.teamfusion.spyglassplus.mixin.exclusive",
   "compatibilityLevel": "JAVA_17",
   "client": [
+    "client.ControlsListWidgetKeyBindingEntryMixin",
     "client.ItemEntityRendererMixin",
     "client.ItemRendererMixin",
-    "client.ModelLoaderMixin"
+    "client.KeyBindingMixin",
+    "client.ModelLoaderMixin",
+    "client.access.ExclusiveKeyBindingAccessor"
   ],
   "injectors": {
     "defaultRequire": 1
